+++ src/main/java/seedu/address/logic/parser/AddPersonCommandParser.java
@@ -11,7 +11,7 @@ import java.util.Set;
 
 import seedu.address.logic.commands.AddPersonCommand;
 import seedu.address.logic.parser.exceptions.ParseException;
-import seedu.address.model.person.Address;
+import seedu.address.model.address.Address;
 import seedu.address.model.person.Email;
 import seedu.address.model.person.Name;
 import seedu.address.model.person.Person;
