+++ src/main/java/seedu/address/model/address/Address.java
@@ -1,13 +1,13 @@
-package seedu.address.model.property;
+package seedu.address.model.address;
 
 import static java.util.Objects.requireNonNull;
 import static seedu.address.commons.util.AppUtil.checkArgument;
 
 /**
- * Represents a Property's address in the address book.
- * Guarantees: immutable; is valid as declared in {@link #isValidPropertyAddress(String)}
+ * Represents a Person's address in the address book.
+ * Guarantees: immutable; is valid as declared in {@link #isValidAddress(String)}
  */
-public class PropertyAddress {
+public class Address {
 
     public static final String MESSAGE_CONSTRAINTS = "Addresses can take any values, and it should not be blank";
 
@@ -24,16 +24,16 @@ public class PropertyAddress {
      *
      * @param address A valid address.
      */
-    public PropertyAddress(String address) {
+    public Address(String address) {
         requireNonNull(address);
-        checkArgument(isValidPropertyAddress(address), MESSAGE_CONSTRAINTS);
+        checkArgument(isValidAddress(address), MESSAGE_CONSTRAINTS);
         value = address;
     }
 
     /**
-     * Returns true if a given string is a valid address.
+     * Returns true if a given string is a valid email.
      */
-    public static boolean isValidPropertyAddress(String test) {
+    public static boolean isValidAddress(String test) {
         return test.matches(VALIDATION_REGEX);
     }
 
@@ -45,8 +45,8 @@ public class PropertyAddress {
     @Override
     public boolean equals(Object other) {
         return other == this // short circuit if same object
-                || (other instanceof PropertyAddress // instanceof handles nulls
-                && value.equals(((PropertyAddress) other).value)); // state check
+                || (other instanceof Address // instanceof handles nulls
+                && value.equals(((Address) other).value)); // state check
     }
 
     @Override
